tests and javadoc

+++ Task_1_1_3/src/main/java/ru/nsu/vostrikov/Number.java
@@ -2,21 +2,36 @@ package ru.nsu.vostrikov;
 
 import java.util.HashMap;
 
+/**
+ * Number class
+ */
 class Number extends Expression {
     private final int value;
 
+    /**
+     * constructor
+     */
     public Number(int value) {
         this.value = value;
     }
 
+    /**
+     * convert to string
+     */
     public String toString() {
         return Integer.toString(value);
     }
 
-    public int calculate(HashMap<String, Integer> dictVars) {
-        return value;
+    /**
+     * calculate variables
+     */
+    public Double calculate(HashMap<String, Double> dictVars) {
+        return (double) value;
     }
 
+    /**
+     * find derivative
+     */
     public Expression derivative(String variable) {
         return new Number(0);
     }
